pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Karolis Vyčius <karolis@vycius.lt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
fastapi-pagination = "^0.12.25"
sqlalchemy = "^2.0.30"
geoalchemy2 = "^0.15.1"
sentry-sdk = { extras = ["fastapi", "sqlalchemy"], version = "^2.5.1" }
sqlean-py = "^3.45.1"
sqlakeyset = "^2.0.1716332987"
gunicorn = "^22.0.0"
uvicorn-worker = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
